# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mini-Lago API (Numaris-native)",
    "version": "2.0.0",
    "description": "Motor de facturación domain-specific para Numaris. Entidades: Customer, Tax, Service (colapsa Plan+Sub+Charge+BM+AddOn), Unit (materializada), EventLog (audit), Invoice, CreditNote.\n\n**Auth**: Bearer en todos los `/api/v1/*` excepto health/openapi/discovery. Callbacks NetSuite usan HMAC-SHA256 (`X-NetSuite-Signature`).\n\n**Idempotency** en POST /invoices y POST /credit_notes vía header `Idempotency-Key` + `metadata.idempotency_key`."
  },
  "servers": [
    { "url": "https://billing2-production.up.railway.app", "description": "Railway production" },
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer" },
      "NetSuiteSignature": { "type": "apiKey", "in": "header", "name": "X-NetSuite-Signature" }
    },
    "schemas": {
...
      "InvoiceWriteable": {
        "type": "object",
        "required": ["service_code"],
        "properties": {
          "service_code": { "type": "string" },
          "period_from": { "type": "string", "format": "date-time", "description": "Si se omite, usa el periodo actual del service." },
          "period_to": { "type": "string", "format": "date-time" },
          "metadata": { "type": "object", "properties": { "idempotency_key": { "type": "string" } } }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sequential_id": { "type": "integer" },
          "number": { "type": ["string", "null"], "description": "Folio fiscal NetSuite." },
          "customer_id": { "type": "string", "format": "uuid" },
          "service_id": { "type": ["string", "null"], "format": "uuid" },
          "status": { "type": "string", "enum": ["calculated", "finalized", "voided"] },
          "external_dispatch_status": { "type": "string", "enum": ["pending", "dispatched", "confirmed", "failed"] },
          "currency": { "type": "string" },
          "fees_amount_cents": { "type": "integer" },
          "taxes_amount_cents": { "type": "integer" },
          "total_amount_cents": { "type": "integer" },
          "period_from": { "type": ["string", "null"], "format": "date-time" },
          "period_to": { "type": ["string", "null"], "format": "date-time" },
          "fees": { "type": "array", "items": { "$ref": "#/components/schemas/Fee" } },
          "units_annex": { "type": "array", "items": { "$ref": "#/components/schemas/UnitsAnnexEntry" } },
          "applied_taxes": { "type": "array", "items": { "type": "object" } },
          "external_invoice": { "type": ["object", "null"] }
        }
      },
      "Fee": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "type": "string", "enum": ["monthly", "setup"] },
          "description": { "type": "string" },
          "units": { "type": "string", "description": "String decimal 4 dígitos." },
          "unit_amount_cents": { "type": "integer" },
          "amount_cents": { "type": "integer" },
          "taxes_amount_cents": { "type": "integer" },
          "total_amount_cents": { "type": "integer" },
          "billed_units_detail": { "type": "array", "items": { "$ref": "#/components/schemas/BilledUnitDetail" } }
        }
      },
      "BilledUnitDetail": {
        "type": "object",
        "properties": {
          "external_id": { "type": "string" },
          "label": { "type": ["string", "null"] },
          "active_from": { "type": "string", "format": "date-time" },
          "active_to": { "type": ["string", "null"], "format": "date-time" },
          "billed_fraction": { "type": "string", "description": "String decimal 4 dígitos (eg. 0.6129)." },
          "amount_cents": { "type": "integer" }
        }
      },
      "UnitsAnnexEntry": {
        "type": "object",
        "properties": {
          "external_id": { "type": "string" },
          "label": { "type": ["string", "null"] },
          "fees": { "type": "array", "items": { "type": "object", "properties": { "kind": { "type": "string" }, "amount_cents": { "type": "integer" } } } }
        }
      },
...
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/api/v1/invoices": {
      "post": {
        "summary": "Calcular invoice del periodo actual del service",
        "tags": ["invoices"],
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "schema": { "type": "string" }, "description": "Debe coincidir con metadata.idempotency_key." }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "invoice": { "$ref": "#/components/schemas/InvoiceWriteable" } } } } } },
        "responses": { "200": { "description": "Invoice", "content": { "application/json": { "schema": { "type": "object", "properties": { "invoice": { "$ref": "#/components/schemas/Invoice" } } } } } } }
      },
      "get": {
        "summary": "List invoices",
        "tags": ["invoices"],
        "parameters": [
          { "name": "customer_external_id", "in": "query", "schema": { "type": "string" } },
          { "name": "service_code", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Lista" } }
      }
    },
    "/api/v1/invoices/{id}": {
      "get": { "summary": "Read invoice", "tags": ["invoices"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Invoice" } } }
    },
    "/api/v1/invoices/{id}/void": {
      "post": { "summary": "Void invoice", "tags": ["invoices"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Invoice voided" } } }
    },
...
  }
}